% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ACTG_test.R
\name{app_ACTG_test}
\alias{app_ACTG_test}
\title{Title   Significance Test for ACTG193A in Real Data Analysis}
\usage{
app_ACTG_test(hc, p_a, type, G1, B, alpha)
}
\arguments{
\item{hc}{the parameter related to the bandwidth}

\item{p_a}{this parameter is gamma of Tn}

\item{type}{the indicator of different hypothesis H_0^type}

\item{G1}{the comparison with Group G1 and Group 4}

\item{B}{the number of bootstrap iteration}

\item{alpha}{the significance level}
}
\value{
the  results for ACTG193A in Real Data Analysis
}
\description{
Title   Significance Test for ACTG193A in Real Data Analysis
}
