% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_twa_test.R
\name{app_twa_test}
\alias{app_twa_test}
\title{Title   Significance Test for TWA in Real Data Analysis}
\usage{
app_twa_test(hc, p_a, type, B, alpha)
}
\arguments{
\item{hc}{the parameter related to the bandwidth}

\item{p_a}{this parameter is gamma of Tn}

\item{type}{the indicator of different hypothesis H_0^type}

\item{B}{the number of bootstrap iteration}

\item{alpha}{the significance level}
}
\value{
the  results for TWA in Real Data Analysis
}
\description{
Title   Significance Test for TWA in Real Data Analysis
}
