% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_w_s1.R
\name{d_w_s1}
\alias{d_w_s1}
\title{Title: Generating the Function Delta(W) in Study 1 under Scenario 1}
\usage{
d_w_s1(w, p, td)
}
\arguments{
\item{w}{the covariates W}

\item{p}{the dimension of W}

\item{td}{the indicator of setting}
}
\value{
the values of Delta(W) and the related arguments
}
\description{
Title: Generating the Function Delta(W) in Study 1 under Scenario 1
}
