% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_pvalue.R
\name{Est_pvalue}
\alias{Est_pvalue}
\title{Title Calculate the empirical p-value}
\usage{
Est_pvalue(n, B, wt, w, y, p_a, hat_res, hat_f, x, hc, alpha)
}
\arguments{
\item{n}{the sample size}

\item{B}{the number of bootstrap iterations}

\item{wt}{the arguments related to respons}

\item{w}{the covariates W}

\item{y}{the observed response#'}

\item{p_a}{the value of parameter gamma in Tn}

\item{hat_res}{the estimator of residual}

\item{hat_f}{the estimator of f(x)}

\item{x}{the covariates x}

\item{hc}{the parameter of the bandwidth}

\item{alpha}{the significance level alpha}
}
\value{
the estimator of P-value
}
\description{
Title Calculate the empirical p-value
}
