% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_res_f.R
\name{Est_res_f}
\alias{Est_res_f}
\title{Title Get the Nonparametric Estimator of e(eta_0) and f(x)}
\usage{
Est_res_f(n, x, hc, y_eta)
}
\arguments{
\item{n}{sample size}

\item{x}{significant variables}

\item{hc}{the parameter related to the bandwidth}

\item{y_eta}{the estimator of y(eta_0)}
}
\value{
the Nonparametric Estimator of e(eta_0) and f(x)
}
\description{
Title Get the Nonparametric Estimator of e(eta_0) and f(x)
}
