% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_Tn.R
\name{Est_Tn}
\alias{Est_Tn}
\title{Title: estiamtion of the test statistic}
\usage{
Est_Tn(hat_res, n, w, p_a, hat_f)
}
\arguments{
\item{hat_res}{the estimated e(hat eta)=hat Y(hat eta)-hat g(X)}

\item{n}{the sample size}

\item{w}{the covariates W}

\item{p_a}{the value of parameter gamma in Tn}

\item{hat_f}{the nonparametric estimator of f(X)}
}
\value{
the value of test statistic
}
\description{
Title: estiamtion of the test statistic
}
