% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_ycate.R
\name{Est_ycate}
\alias{Est_ycate}
\title{Title Estimate hat Y(eta) in Study 1}
\usage{
Est_ycate(y, D, w, wt)
}
\arguments{
\item{y}{the observed response}

\item{D}{the treatment variable}

\item{w}{the covriates W}

\item{wt}{the related arguments}
}
\value{
the value of estimated y(eta)
}
\description{
Title Estimate hat Y(eta) in Study 1
}
