% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_ytwa.R
\name{Est_ytwa}
\alias{Est_ytwa}
\title{Title Estiamte the Y(eta) in TWA application}
\usage{
Est_ytwa(y, D, w, wt)
}
\arguments{
\item{y}{the observed response}

\item{D}{the treatment variable}

\item{w}{the covriates W}

\item{wt}{the related arguments}
}
\value{
the value of estimated y(eta)
}
\description{
Title Estiamte the Y(eta) in TWA application
}
