% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_x_s2.R
\name{g_x_s2}
\alias{g_x_s2}
\title{Title: generating the function g(X)}
\usage{
g_x_s2(w, tg)
}
\arguments{
\item{w}{the covariates W}

\item{tg}{the indicator of setting}
}
\value{
the values of g(X) and the related arguments
}
\description{
Title: generating the function g(X)
}
