% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gakernel_4.R
\name{Gakernel_4}
\alias{Gakernel_4}
\title{Title high-order Guassian kernel}
\usage{
Gakernel_4(x)
}
\arguments{
\item{x}{the covariate}
}
\value{
the nonparametric estimator
}
\description{
Title high-order Guassian kernel
}
