% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_D.R
\name{Gen_D}
\alias{Gen_D}
\title{Title: generating the treatment variable D}
\usage{
Gen_D(rep1, arg, n, p, w)
}
\arguments{
\item{rep1}{the indicator of rep1-th simulation run}

\item{arg}{the indicator of the Scenario "s1" or "s2"}

\item{n}{the sample size}

\item{p}{the dimension of W}

\item{w}{the covariates W}
}
\value{
the values of treatment D
}
\description{
Title: generating the treatment variable D
}
