% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_W.R
\name{Gen_W}
\alias{Gen_W}
\title{Title generating the covariates W}
\usage{
Gen_W(rep1, n, p)
}
\arguments{
\item{rep1}{the indicator of rep1-th simulation run}

\item{n}{the sample size}

\item{p}{the dimension of W}
}
\value{
the values of covariates W
}
\description{
Title generating the covariates W
}
