% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_Ycate.R
\name{Gen_Ycate}
\alias{Gen_Ycate}
\title{Title: generating the observed response Y and the related arguments}
\usage{
Gen_Ycate(rep1, arg, n, Type, p, w, D, a)
}
\arguments{
\item{rep1}{the indicator of rep1-th simulation run}

\item{arg}{the Indicator of Scenario in Study1}

\item{n}{the sample size}

\item{Type}{a (i,j,k) vector which is the indicator of setting}

\item{p}{the dimension of W}

\item{w}{the covariates W}

\item{D}{the treatment variable D}

\item{a}{the dispersion parameter a}
}
\value{
the values of response Y and the related arguments
}
\description{
Title: generating the observed response Y and the related arguments
}
