% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gen_Ymiss.R
\name{Gen_Ymiss}
\alias{Gen_Ymiss}
\title{Title: generating the observed response Y and the related arguments}
\usage{
Gen_Ymiss(rep1, n, Type, p, w, D, a)
}
\arguments{
\item{rep1}{the indicator of rep1-th simulation run}

\item{n}{the sample size}

\item{Type}{a (i,j) vector which is the indicator of setting}

\item{p}{the dimension of W}

\item{w}{the covariates W}

\item{D}{the missing variable D}

\item{a}{the dispersion parameter a}
}
\value{
the values of response Y and the related arguments
}
\description{
Title: generating the observed response Y and the related arguments
}
