% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_w_s2.R
\name{h_w_s2}
\alias{h_w_s2}
\title{Title:generating the function h(W)}
\usage{
h_w_s2(w, th, p)
}
\arguments{
\item{w}{the covariates W}

\item{th}{the indicator of setting}

\item{p}{the dimension of W}
}
\value{
the values of h(W)
}
\description{
Title:generating the function h(W)
}
