% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_study2.R
\name{process_study2}
\alias{process_study2}
\title{Title Simulation results for Missing Data in rep1-th simulation run}
\usage{
process_study2(rep1, n, B, hc, p_a, p, q, alpha, Type_y, disp_a)
}
\arguments{
\item{rep1}{the indicator of rep1-th simulation run}

\item{n}{the sample size}

\item{B}{the number of bootstrap iteration#'}

\item{hc}{the parameter related to the bandwidth#'}

\item{p_a}{this parameter is gamma of Tn#'}

\item{p}{the dimension of W}

\item{q}{the dimension of x}

\item{alpha}{the significance level}

\item{Type_y}{the indicator of the model type (i,j,k)}

\item{disp_a}{the dispersion parameter a}
}
\value{
the simulation results of rep1-th simulation run
}
\description{
Title Simulation results for Missing Data in rep1-th simulation run
}
