% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tempwork.R
\name{tempwork}
\alias{tempwork}
\title{Data about Temporary Work Agency (TWA) Assignments}
\description{
The data is publicly available at http://qed.econ.queensu.ca/jae/2008-v23.3/ichino-mealli-nannicini.
}
\keyword{dataset}
