% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_miss_study2.R
\name{test_miss_study2}
\alias{test_miss_study2}
\title{Title Simulation of Significance Test for Missing Data in Study 2}
\usage{
test_miss_study2(rep_num, n, B, hc, p_a, p, q, alpha, Type_y)
}
\arguments{
\item{rep_num}{the number of simulation iteration}

\item{n}{the sample size}

\item{B}{the number of bootstrap iteration}

\item{hc}{the parameter related to the bandwidth}

\item{p_a}{this parameter is gamma of Tn}

\item{p}{the dimension of W}

\item{q}{the dimension of x}

\item{alpha}{the significance level}

\item{Type_y}{the indicator of the model type (i,j)}
}
\value{
the simulation results for Missing Data in Study 2
}
\description{
Title Simulation of Significance Test for Missing Data in Study 2
}
